pyproject.toml
[tool.poetry]
name = "robiemon_server"
version = "0.0.1"
description = ""
authors = ["endaaman"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["uvicorn"], version = "^0.109.0"}
sqlalchemy = "^2.0.25"
pydantic = "^1"
torch = "^2.1.2"
timm = "^0.9.12"
opencv-python-headless = "^4.9.0.80"
grad-cam = "^1.5.0"
numpy = "^1.26.3"
python-multipart = "^0.0.6"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
pillow = "^10.2.0"
sqlite-web = "^0.6.3"
taskipy = "^1.12.2"
gunicorn = "^22.0.0"
pandas = "^2.2.2"
polars = "^0.20.22"
redis = "^5.0.3"
aioredis = "^2.0.1"
xlsxwriter = "^3.2.0"
openpyxl = "^3.1.2"
watchdog = "^4.0.0"

[tool.poetry.group.dev.dependencies]
jedi-language-server = "^0.41.2"

[build-system]

[tool.taskipy.tasks]
# gunicorn
# dev = "gunicorn robiemon_server:app -w 4 --bind 0.0.0.0:3000 --reload"
dev = "uvicorn robiemon_server:app --reload --host 0.0.0.0 --port 3000"
prod = "uvicorn robiemon_server:app --host 0.0.0.0 --port 3000"
admin = "sqlite_web --no-browser data/server.db"
